style: touched up style of addtional details screen
